+++ changelog.md
@@ -1,6 +1,7 @@
 0.2.0 (2018-06-25)
 ==================
 
+* #33: Added `Request.accept()` method.
 * #35: Added `Request.type` and `Response.type`.
 * #36: Added `Request.query`.
 * #37: `Response.body` now has type `any`.
@@ -12,16 +13,19 @@
 
 * Set `script` and `types` correctly in `package.json`.
 
+
 0.1.1 (2018-06-24)
 ==================
 
 * Fixed npm package distribution. Was shipping the wrong files.
+
 
 0.1.0 (2018-06-24)
 ==================
 
 * Created `Request`, `Response`, `Application`, `Context`, `Headers` classes.
 * Basic framework works
+
 
 0.0.1 (2018-06-23)
 ==================